+++ src/scenes/stage5.js
@@ -1,0 +1,5 @@
+const mole = () => {
+  console.log('ddddddddddddddddddddd')
+};
+
+export default mole;
